Convert js BigInt to regular num if possible
