+++ sqlite3/lib/src/wasm/big_int.dart
@@ -2,10 +2,16 @@ import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 
 @JS('BigInt')
-external Object Function(String string) get _jsBigInt;
+external Object _jsBigInt(String s);
 
 @JS('Number')
-external int Function(Object obj) get _jsNumber;
+external int _jsNumber(Object obj);
+
+@JS('eval')
+external Object _eval(String s);
+
+bool Function(Object, Object) jsLeq =
+    _eval('(a, b) => a <= b') as bool Function(Object, Object);
 
 Object bigIntToJs(BigInt i) => _jsBigInt(i.toString());
 BigInt jsToBigInt(Object jsObject) =>
