+++ sqlite3/test/wasm/common_database_test.dart
@@ -1,0 +1,6 @@
+import '../common/database.dart';
+import 'utils.dart';
+
+void main() {
+  testDatabase(loadSqlite3);
+}
